+++ src/lib/krb4/realmofhost.c
@@ -105,7 +105,7 @@ krb_realmofhost(host)
 		/* Upper-case realm */
 		for (cp = ret_realm; *cp; cp++)
 			if (islower((int) (*cp)))
-				*cp = toupper(*cp);
+				*cp = toupper((int) *cp);
 	} else {
 		krb_get_lrealm(ret_realm, 1);
 	}
